requirements.txt
flask
flask-cors
python-dotenv
google-generativeai>=0.7.0
Pillow